kernel:shim-eth-vlan: consolidated separate files into shim-eth-vlan, refined shim-eth-qdisc; rate-limiting seems to be working
