+++ linux/net/rina/ipcps/Makefile
@@ -10,6 +10,5 @@ ccflags-y += -Wtype-limits
 
 obj-$(CONFIG_RINA_NORMAL_IPCP)    += normal-ipcp.o
 obj-$(CONFIG_RINA_SHIM_ETH_VLAN)  += shim-eth-vlan.o
-shim-eth-vlan-y := shim-eth-qdisc.o		
 obj-$(CONFIG_RINA_SHIM_TCP_UDP)   += shim-tcp-udp.o
 obj-$(CONFIG_RINA_SHIM_HV)        += shim-hv.o
