--- linux/net/rina/ipcps/shim-eth-qdisc.h
@@ -1,33 +1,0 @@
-/*
- * Qdisc for the shim-eth-vlan (will enable all ports the ipcp uses)
- *
- *   Eduard Grasa <eduard.grasa@i2cat.net>
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
- */
-
-#ifndef SHIM_ETH_QDISC_H
-#define SHIM_ETH_QDISC_H
-
-#include <linux/netdevice.h>
-
-#include "common.h"
-
-struct Qdisc * shim_eth_qdisc_alloc(ipc_process_id_t id,
-			            struct net_device * dev);
-
-void shim_eth_qdisc_destroy(struct net_device * dev);
-
-#endif
