+++ linux/net/rina/ipcps/shim-eth-vlan.c
@@ -27,9 +27,11 @@
 #include <linux/string.h>
 #include <linux/list.h>
 #include <linux/netdevice.h>
+#include <linux/if.h>
 #include <linux/if_packet.h>
 #include <linux/workqueue.h>
 #include <linux/notifier.h>
+#include <net/pkt_sched.h>
 #include <net/sch_generic.h>
 
 #define SHIM_NAME   "shim-eth-vlan"
@@ -37,8 +39,6 @@
 #define RINA_PREFIX SHIM_NAME
 
 #include "logs.h"
-#include "shim-eth-vlan.h"
-#include "shim-eth-qdisc.h"
 #include "kipcm.h"
 #include "debug.h"
 #include "utils.h"
@@ -738,7 +738,7 @@ static int eth_vlan_application_unregister(struct ipcp_instance_data * data,
         return 0;
 }
 
-void eth_vlan_enable_write_all(ipc_process_id_t id)
+static void eth_vlan_enable_write_all(ipc_process_id_t id)
 {
 	struct ipcp_instance_data * ipcp_data;
 	struct shim_eth_flow 	  * flow;
@@ -759,7 +759,145 @@ void eth_vlan_enable_write_all(ipc_process_id_t id)
 	}
 	spin_unlock_irqrestore(&ipcp_data->lock, flags);
 }
-EXPORT_SYMBOL(eth_vlan_enable_write_all);
+
+/*
+ * Private data for a rina_shim_eth scheduler containing:
+ * (only supports a single flow for the moment)
+ *	- ipcp_id id of the shim-eth IPCP that is using this qdisc
+ * 	- queue max_size
+ * 	- queue enable_thres (size at which the qdisc will enable N+1 ports)
+ * 	- the queue
+ */
+struct shim_eth_qdisc_priv {
+	ipc_process_id_t    ipcp_id;
+	uint16_t            q_max_size;
+	uint16_t            q_enable_thres;
+};
+
+static int shim_eth_qdisc_enqueue(struct sk_buff *skb, struct Qdisc *qdisc)
+{
+	struct shim_eth_qdisc_priv *priv = qdisc_priv(qdisc);
+
+	LOG_DBG("shim-eth-enqueue called; current size is %u", qdisc->q.qlen);
+	if (skb_queue_len(&qdisc->q) < priv->q_max_size) {
+		return __qdisc_enqueue_tail(skb, qdisc, &qdisc->q);
+	}
+
+	return qdisc_drop(skb, qdisc);
+}
+
+static struct sk_buff *shim_eth_qdisc_dequeue(struct Qdisc *qdisc)
+{
+	struct shim_eth_qdisc_priv *priv = qdisc_priv(qdisc);
+
+	LOG_DBG("shim-eth-dequeue called; current size is %u", qdisc->q.qlen);
+
+	if (skb_queue_len(&qdisc->q) > 0) {
+		struct sk_buff *skb = __qdisc_dequeue_head(qdisc, &qdisc->q);
+		if (skb_queue_len(&qdisc->q) == priv->q_enable_thres) {
+			LOG_INFO("Enabling all ports");
+			eth_vlan_enable_write_all(priv->ipcp_id);
+		}
+
+		return skb;
+	}
+
+	return NULL;
+}
+
+static struct sk_buff * shim_eth_qdisc_peek(struct Qdisc *qdisc)
+{
+	return skb_peek(&qdisc->q);
+}
+
+static int shim_eth_qdisc_init(struct Qdisc *qdisc, struct nlattr *opt)
+{
+	struct shim_eth_qdisc_priv * priv;
+
+	if (!opt) {
+		return 0;
+	}
+
+	priv = qdisc_priv(qdisc);
+	priv->ipcp_id = opt->nla_len;
+	priv->q_max_size = 20;
+	priv->q_enable_thres = 5;
+	skb_queue_head_init(&qdisc->q);
+
+	LOG_INFO("shim-eth-init: ipcp id %u, max size: %u, enable thres: %u",
+		 priv->ipcp_id, priv->q_max_size, priv->q_enable_thres);
+
+	return 0;
+}
+
+static void shim_eth_qdisc_reset(struct Qdisc *qdisc)
+{
+	__qdisc_reset_queue(qdisc, &qdisc->q);
+
+	qdisc->qstats.backlog = 0;
+	qdisc->q.qlen = 0;
+}
+
+static struct Qdisc_ops shim_eth_qdisc_ops __read_mostly = {
+	.id		=	"rina_shim_eth",
+	.priv_size	=	sizeof(struct shim_eth_qdisc_priv),
+	.enqueue	=	shim_eth_qdisc_enqueue,
+	.dequeue	=	shim_eth_qdisc_dequeue,
+	.peek		=	shim_eth_qdisc_peek,
+	.init		=	shim_eth_qdisc_init,
+	.reset		=	shim_eth_qdisc_reset,
+	.dump		=	NULL,
+	.owner		=	THIS_MODULE,
+};
+
+static int eth_vlan_update_qdisc(struct ipcp_instance_data * data)
+{
+	struct Qdisc * sch;
+	struct nlattr  arg;
+
+	sch = qdisc_create_dflt(data->dev->_tx, &shim_eth_qdisc_ops, 0);
+	if (!sch) {
+		LOG_ERR("Problems creating shim-eth-qdisc");
+		return -1;
+	}
+
+	arg.nla_len = data->id;
+	if (shim_eth_qdisc_ops.init(sch, &arg)) {
+		LOG_ERR("Problems initializing shim-eth-qdisc");
+		qdisc_destroy(sch);
+		return -1;
+	}
+
+	data->old_qdisc = data->dev->qdisc;
+
+	if (data->dev->flags & IFF_UP)
+		dev_deactivate(data->dev);
+
+	dev_graft_qdisc(netdev_get_tx_queue(data->dev, 0), sch);
+	data->dev->qdisc = sch;
+
+	if (data->dev->flags & IFF_UP)
+		dev_activate(data->dev);
+
+
+	return 0;
+}
+
+static void eth_vlan_restore_qdisc(struct ipcp_instance_data * data)
+{
+	struct Qdisc * sch = data->dev->qdisc;
+
+	if (data->dev->flags & IFF_UP)
+		dev_deactivate(data->dev);
+
+	dev_graft_qdisc(netdev_get_tx_queue(data->dev, 0), data->old_qdisc);
+	data->dev->qdisc = data->old_qdisc;
+	data->old_qdisc = NULL;
+	qdisc_destroy(sch);
+
+	if (data->dev->flags & IFF_UP)
+		dev_activate(data->dev);
+}
 
 static int eth_vlan_sdu_write(struct ipcp_instance_data * data,
                               port_id_t                   id,
@@ -1328,11 +1466,8 @@ static int eth_vlan_assign_to_dif(struct ipcp_instance_data * data,
         }
 
         /* Modfy qdisc by our own */
-        data->old_qdisc = data->dev->qdisc;
-        data->dev->qdisc = shim_eth_qdisc_alloc(data->id, data->dev);
-        if (!data->dev->qdisc) {
+        if (eth_vlan_update_qdisc(data)) {
         	LOG_ERR("Problems creating queue discipline");
-        	data->dev->qdisc = data->old_qdisc;
         	read_unlock(&dev_base_lock);
         	name_destroy(data->dif_name);
         	data->dif_name = NULL;
@@ -1408,8 +1543,7 @@ static int eth_vlan_update_dif_config(struct ipcp_instance_data * data,
                 }
         }
 
-        shim_eth_qdisc_destroy(data->dev);
-        data->dev->qdisc = data->old_qdisc;
+        eth_vlan_restore_qdisc(data);
 
         dev_remove_pack(data->eth_vlan_packet_type);
         /* Remove from list */
@@ -1444,11 +1578,8 @@ static int eth_vlan_update_dif_config(struct ipcp_instance_data * data,
         }
 
         /* Modfy qdisc by our own */
-        data->old_qdisc = data->dev->qdisc;
-        data->dev->qdisc = shim_eth_qdisc_alloc(data->id, data->dev);
-        if (!data->dev->qdisc) {
+        if (eth_vlan_update_qdisc(data)) {
         	LOG_ERR("Problems creating queue discipline");
-        	data->dev->qdisc = data->old_qdisc;
         	read_unlock(&dev_base_lock);
         	name_destroy(data->dif_name);
         	data->dif_name = NULL;
@@ -1768,8 +1899,7 @@ static int eth_vlan_destroy(struct ipcp_factory_data * data,
                         }
 
                         /* Restore old qdisc */
-                        shim_eth_qdisc_destroy(pos->dev);
-                        pos->dev->qdisc = pos->old_qdisc;
+                        eth_vlan_restore_qdisc(pos);
 
                         /* Remove packet handler if there is one */
                         if (pos->eth_vlan_packet_type->dev)
