--- linux/net/rina/ipcps/shim-eth-vlan.h
@@ -1,31 +1,0 @@
-/*
- * Shim IPC Process over Ethernet (using VLANs)
- *
- *   Francesco Salvestrini <f.salvestrini@nextworks.it>
- *   Sander Vrijders       <sander.vrijders@intec.ugent.be>
- *   Miquel Tarzan         <miquel.tarzan@i2cat.net>
- *   Leonardo Bergesio     <leonardo.bergesio@i2cat.net>
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
- */
-
-#ifndef SHIM_ETH_VLAN_H
-#define SHIM_ETH_VLAN_H
-
-#include "common.h"
-
-void eth_vlan_enable_write_all(ipc_process_id_t id);
-
-#endif
